REF: Components isolated from BlueComponents
